+++ Pruebas/script_clientes_paquetes.sql
@@ -9,7 +9,7 @@ BEGIN
     CREATE TABLE compra_cliente (
         id INT IDENTITY(1,1) PRIMARY KEY,
         id_usuario INT,
-        id_paquete bigint,
+        id_paquete INT,
 
 		FOREIGN KEY (id_usuario) REFERENCES usuarios(id),
 		FOREIGN KEY (id_paquete) REFERENCES paquetes(id_paquete)
